requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Werkzeug==2.3.7
Pillow>=10.0.1
PyPDF2==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
qrcode==7.4.2
cryptography==41.0.5
requests==2.31.0
email-validator==2.1.0
